+++ go/lib/drkey/drkeydbsqlite/BUILD.bazel
@@ -3,9 +3,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "common.go",
         "lvl1db.go",
         "lvl2db.go",
-        "schema.go",
     ],
     importpath = "github.com/scionproto/scion/go/lib/drkey/drkeydbsqlite",
     visibility = ["//visibility:public"],
