--- go/lib/drkey/drkeydbsqlite/schema.go
@@ -1,55 +1,0 @@
-// Copyright 2019 ETH Zurich
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package drkeydbsqlite
-
-const (
-	// SchemaVersion is the version of the SQLite schema understood by this backend.
-	// Whenever changes to the schema are made, this version number should be increased
-	// to prevent data corruption between incompatible database schemas.
-	Lvl1SchemaVersion = 1
-	// Schema is the SQLite database layout.
-	Lvl1Schema = `
-	CREATE TABLE DRKeyLvl1 (
-		SrcIsdID 	INTEGER NOT NULL,
-		SrcAsID 	INTEGER NOT NULL,
-		DstIsdID 	INTEGER NOT NULL,
-		DstAsID 	INTEGER NOT NULL,
-		EpochBegin 	INTEGER NOT NULL,
-		EpochEnd 	INTEGER NOT NULL,
-		Key 		TEXT NOT NULL,
-		PRIMARY KEY (SrcIsdID, SrcAsID, DstIsdID, DstAsID, EpochBegin)
-	);`
-
-	// SchemaVersion is the version of the SQLite schema understood by this backend.
-	// Whenever changes to the schema are made, this version number should be increased
-	// to prevent data corruption between incompatible database schemas.
-	Lvl2SchemaVersion = 1
-	// Schema is the SQLite database layout.
-	Lvl2Schema = `
-	CREATE TABLE DRKeyLvl2 (
-		Protocol	TEXT NOT NULL,
-		Type		INTEGER NOT NULL,
-		SrcIsdID 	INTEGER NOT NULL,
-		SrcAsID 	INTEGER NOT NULL,
-		DstIsdID 	INTEGER NOT NULL,
-		DstAsID 	INTEGER NOT NULL,
-		SrcHostIP 	TEXT,
-        DstHostIP	TEXT,
-        EpochBegin  INTEGER NOT NULL,
-        EpochEnd    INTEGER NOT NULL,
-		Key 		TEXT NOT NULL,
-		PRIMARY KEY (Protocol, Type, SrcIsdID, SrcAsID, DstIsdID, DstAsID, SrcHostIP, DstHostIP, EpochBegin)
-	);`
-)
